requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# aqua/market_data/alpaca.py: 10
# aqua/market_data/polygon.py: 10
aiohttp == 3.7.4.post0

# aqua/market_data/ibkr.py: 7,8,9
ibapi == 9.81.1.post1

# aqua/market_data/alpaca.py: 11
# aqua/market_data/market_data_interface.py: 8
# aqua/market_data/polygon.py: 11
# aqua/security/option.py: 7
pandas == 1.3.2

# aqua/market_data/alpaca.py: 12
# aqua/market_data/ibkr.py: 10
# aqua/market_data/polygon.py: 12
python_dotenv == 0.19.0
